changes to work with shorti
