+++ 500g/BUKI/platformio.ini
@@ -8,7 +8,7 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:teensy41]
-platform = teensy
-board = teensy41
+[env:LGT8F328P]
+platform = lgt8f
+board = LGT8F328P
 framework = arduino
